refactor: replace polling-with-timeouts with reactive DB listener for reminder metadata transfer
The original approach in transferReminderMetadataFromDraftToThread used a
fixed retry schedule (1.5s + 1.5s + 10s + 10s + 15s + 22s ≈ 60s) to poll
for the sent message to appear in the local database. This had two problems:

1. It fails silently if the user is offline or sync takes longer than ~60s,
   permanently losing the reminder they set on the draft.
2. The async function was called without .catch() in the store, turning a
   timeout failure into an unhandled promise rejection reported to Sentry
   (Sentry issue MAILSPRING-CLIENT-J, 93 users, 465 events).

The fix replaces the polling approach with the same reactive database pattern
the rest of the app uses:

- When draftDeliverySucceeded fires, the {headerMessageId, accountId} pair
  is added to an in-memory _pendingMetadataTransfers map, and we do a single
  optimistic DB check in case the message is already there.

- _onDatabaseChanged now also handles Message change records. When any Message
  lands in the database, we check whether its headerMessageId is in the pending
  map. If so, we perform the transfer immediately and remove from the map.

This means:
- If the sync engine returns the message in 100ms, the transfer happens in 100ms.
- If it takes 10 minutes (offline, slow server), it still works correctly.
- No arbitrary timeouts, no unhandled rejections, no Sentry noise.

The transfer logic (was in send-reminders-utils.ts) is now inlined into
_completePendingTransfer in the store, and the no-longer-needed
transferReminderMetadataFromDraftToThread / findMessage functions are removed.

Fixes MAILSPRING-CLIENT-J
